Progress on platform movement.  Introduced SharedPoints to link positions of objects together.

+++ Source/Geometry/FPLong.cs
@@ -335,7 +335,7 @@ namespace Amphibian.Geometry
 
         public override string ToString ()
         {
-            return (_raw >> ShiftBy) + "+" + ((_raw & SubMask) / OneL);
+            return (_raw >> ShiftBy) + "+" + ((float)(_raw & SubMask) / OneL);
         }
 
         #endregion
